Completed a working version of BH in BHTree
added a finder function to find corresponding node for particle (recursive) and force calc based on barnes-hut algorithm, also recursive. it might be a good idea to create paralelization for this depending on where the bottlenecks end up being. Attempted paralelization of force AB but didnt integrate with engine Processes. see lines 209-220.